move fetch_jobs
